fix: revue de code — clôture, départs employés, ratios et cookies
- StageReportService : exclusion des départs non confirmés (auto_closed / claimed) du ratio des journées complètes, alignement de la pause et fallback horaire entreprise
- AutoCheckout : rattrapage sur 7 jours des journées antérieures non clôturées et eager loading du stage
- PresenceService : calcul effectif des départs anticipés pour les employés et cast entier anti-dépréciation
- WorkScheduleResolver : cast entier pour diffInMinutes (compatibilité PHP 8.x/9)
- QrPointageController : log des erreurs du préflight et nettoyage du cookie à la révocation
- Ajout de PLAN_POINTAGE_QR.html et exclusion de .claude

+++ app/Services/WorkScheduleResolver.php
@@ -69,7 +69,10 @@ class WorkScheduleResolver
      */
     public function workedMinutes(?Stage $stage, Carbon $checkIn, Carbon $checkOut): int
     {
-        $presence = max(0, $checkIn->diffInMinutes($checkOut));
+        // diffInMinutes renvoie un flottant : sans conversion explicite, le
+        // type de retour int déclenche une déprécation à chaque pointage, et
+        // une TypeError en PHP 9.
+        $presence = (int) max(0, $checkIn->diffInMinutes($checkOut));
         $break    = $this->forStage($stage, $checkIn)['break_minutes'];
 
         return $presence > $break ? $presence - $break : $presence;
